added builtin tokenize funcitonality & tests (#14)
* added builtin tokenize funcitonality & tests

* added vendor whoops

* renamed tokenize to match python sdk, small test change

* updated tokenizer library

+++ tokens.go
@@ -7,5 +7,5 @@ type TokenizeOptions struct {
 
 type TokenizeResponse struct {
 	// The tokens.
-	Tokens []uint `json:"tokens"`
+	Tokens []int64 `json:"tokens"`
 }
